Revert "Create GetListFromFile method and change flow Update method"
This reverts commit fede0904f5954286aa1d022e20103a1582a82f53.
